remove delayed messages from batches (#416)
This allows the ingestor to build batches instead of the block builder
by removing the dependency of batches on settlement chain blocks.

Under the new architecture, the ingestor will send arbitrum batches
to the slotter over the sequencing channel and delayed messages to the slotter
over the settlement channel. The slotter memcopies these structures into a slot
object and sends them to the mchain which persists them to disk.

This is a breaking change - deposits inside a slot are now ordered
after sequencer transactions instead of before.

If we really wanted to avoid breaking compatibility with existing
rollups while removing the dependency of batches on the settlement
chain, we can potentially split the batch in two - an empty batch
followed by a regular one - to force delayed messages to be sequenced
before sequencer ones. imo this approach is more complicated and not worth
it though.

--------- V1.5 notes ---------

This enables TEE proofs to be generated for batches by the ingestor,
something along the lines of:

[(batch 1, block hash 1, proof 1), (batch 2, block hash 2, proof 2), ...]

where proof 2 = TEE signature (block hash 1, block hash 2, batch 2)
which proves that block 1 is the parent of block 2 and batch 2 is
produced from block 2.

There is a total overhead of 32 bytes for the (batch, block, proof) tuple
compared to just sending (batch, block hash): 32 bytes for the TEE proof hash.

We always need to include the block hash for reorg detection (it is
stored in the mchain db)

*If the latency to generate a TEE proof is significant, which I
suspect it will not be, the ingestor can send (batch data, block hash)
immediately to the metatranslator and follow up with the
TEE proof in a future message when it is ready.

Similarly TEE proofs can be generated for DelayedMessages by the
ingestor, something along of the lines of:

[(Vec<DelayedMessage>, block hash 1, proof 1), ...]

where proof 2 = TEE signature (block hash 1, block hash 2, delayed messages 2)

Basically: TEE proofs can be generated independently and in parallel for sequencing
and settlement chain blocks.

These TEE proofs replace the verifier component in withdrawals 1.5:
raw batches and Vec<DelayedMessages> can be fed directly into the nitro
block verifier. Or to be precise: it removes the dependency of the
verifier component on ingestor block data! The verifier now does not
require any external data from the settlement or sequencing chains.

----- V1.5 superingestor  -------

The ingestor crate should be separated out from the metabased translator into
a separate program, the superingestor. One superingestor runs per a
sequencing or settlement chain. The superingestor hosts a websocket
server that provides an API to metabased translator applications to
1) subscribe to a stream of Batch/DelayedMessage updates as each new
   block is ingested
2) request historic Batch/DelayedMessage message from a db persisted to
   disk

It also contains an admin endpoint to add a new address to monitor for
events to the list of Bridge/Inbox addreses whenever a new rollup is
deployed.

The superingestor persists sequencing (batch, block hash, TEE proof) tuples
and settlement (delaymessages, block hash, TEE proof) tuples to disk in a
rocksdb database for each rollup contract that it is configured to
monitor. The same fetched block data is reused for all the rollups -
this allows us to keep the alchemy bill low and enables the assertion
poster service to generate a TEE proof to post without needing to fetch
any block data.

The TEE proof has an overhead of 32 bytes per rollup per block, which is
very tractable (64 bytes total for settlement + sequencing chains).

------- public replica instructions -----

To run a replica node for an appchain, users need to run the following
components:

1) Sequencer superingestor
2) Settlement superingestor
3) Metabased translator
4) Nitro node

This setup automatically generates and challenges the TEE state root on
the settlement chain if fraud is detected.

Users can optionally run a replica node for the sequencer chain eg exo
to challenge TEE proofs of the latest sequencing block posted to the
poster contract on the settlement chain. They do not need to run said
replica node because it inherits its security from the users of all
appchains that use the chain for sequencing.

------ sequencing chain -> settlement chain TEE messenger ------

the sequencing -> settlement TEE messenger has 2 steps

1) ascend up the proof ladder of parent chains. Each step up the ladder requires
   running the execution client used to derive the child chain in order to
   derive past the end of the finalized state root.

2A) jump from a l1 chain eg holesky to another l1 eg sepolia. This
   involves verifying that the **source** l1 chain block is signed by
   the **sync committe**. Note that the destination chain is unused for
   this step except to pick a timestamp to anchor the latest source
   chain block used for derivation to, or more precisely: to verify that
   said timestamp is in the past and therefore said block exists and
   the proof is challengeable.

OR:

2B) descend down the ladder of child chains - no need to run any software for this
either A) examine the block header if the parent origin block hash is included
in it B) use precompile ie L1BLock to get the parent hash or C) send a message from
the parent->child chain that includes the parent hash in it.

A jump is necessary when there is no common ancestor in between the
sequencing and settlement chains. A jump can be used in all cases and
allows the destination chain to be ignored.