various improvements. in terms of breaking changes:
added the mbt flag --allowed_settlement_addresses for dream

block.number now returns the sequencing block number of the latest
transaction sequenced from the sequencing chain instead of the mchain
block number of the latest mchain block that includes a tx from the
sequencing chain. This matches the way regular arbitrum chains are
supported to behave as per
https://docs.arbitrum.io/build-decentralized-apps/arbitrum-vs-ethereum/solidity-support.

Note that the sequencing chain block number is used instead of the
settlement chain one to allow the ingestor to build the batch instead of
the slotter.

Additionally note that block.number and block.timestamp are not updated
when delayed messages are sequenced in order to match the behavior of
the default nitro sequencer - I think it only inserts advance l1 block
number and advance timestamp messages for regular transactions and not
delayed inbox ones. The matches the current behavior of the mbt.

Block builder metrics have been removed as it now builds blocks out of
order - slotter metrics should be used instead to track the chain state.