disable nitro fork by default
The syndicate fork is disabled at the sequencer contract level by blocking unsigned
transactions from being emitted. The enableSyndicateFork function can be
called when we are ready to enable the fork.

While the fork is disabled, the appchain will be compatible with both
forked and upstream nitro versions.

Additional changes: a max user tx limit of 100 per block is enforced to
ensure that transations have enough gas to be included in the block -
on average, 320,000 gas is available per tx if the block is filled.

Transactions will automatically be split into new appchain blocks when
the 256 kb max l2 message size limit is exceeded.