update permission module interface
Previously, signed transactions were represented verbatim, unsigned ones
as the unsigned tx type prefix, and compressed ones as null. Now
unsigned transactions include the full tx data so that the sequencing module
can use it as part of the isAllowed() check. Additionally, signed transactions
include the signed transaction type prefix to make them easier to detect,
prevent collisions with the other tx types, and allow for new tx types
with a different prefix byte to be added in the future. Finally
compressed transactions are represented as the compressed tx prefix byte
instead of null to prevent reverts when checking the tx type.