revisions
rename preComputeDiminishingFactors to computeDiminishingFactors

make it mandatory to call computeDiminishingFactors() prior to claiming
- that way, the claim function does not use a large amount of gas

use a separate EmissionsReceiver contract to track the appchain reward address
so that multiple pools can access this

do not sort aggregation data to save gas

ignore appchains with no gas usage in the epoch to save gas - this
should not affect the reward formulas

add a function getAppchainInfo() that returns a list of gas usage for
appchains - this is more efficient than fetching gas usage and chain id
separately. getAppchainInfo supports pagination of results.